requirements.txt
langchain==0.1.11
langchain-openai==0.1.1
langchain-community==0.0.29
langchain-core==0.1.30
langchain-experimental==0.0.54
langgraph==0.0.43
python-dotenv==1.0.1
openai